Added CS50 Library -- developed for Harvard's CS50x course
This project makes use of the CS50 library, which was developed as part of Harvard University's CS50x: Introduction to Computer Science course. The source files for the CS50 library are publicly available and can be found at https://cs50.harvard.edu. Special thanks to the instructors and contributors of CS50x for making this valuable resource available to the programming community.